+++ src/lib/components/FieldCheckboxes/Checkbox.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import T from 'prop-types';
 
+import * as propUtils from '../../utils/propTypes';
+
 // handle changing array value in redux form
 export const handleChange = ({ value, values, onChange }) => (event) => {
   const newValues = [...values];
@@ -20,66 +22,94 @@ export const handleChange = ({ value, values, onChange }) => (event) => {
 };
 
 // individual checkbox, applies state-template styling
-const Checkbox = ({
-  input, value, label, variant, disabled,
-}) => {
-  const values = Array.isArray(input.value) ? input.value : [input.value];
-  const checked = values.includes(value);
-  const onChange = handleChange({ value, values, onChange: input.onChange });
-
-  const inputProps = {
+const Checkbox = (props) => {
+  const {
+    name,
+    value,
+    option,
+    variant,
     disabled,
-    checked,
     onChange,
-    type: 'checkbox',
-    name: input.name,
-  };
+    'aria-invalid': ariaInvalid,
+    'aria-describedby': ariaDescribedby,
+  } = props;
 
-  const id = `${input.name}-${value}`;
+  const values = Array.isArray(value) ? value : [value];
+
+  const id = `${name}-${option.value}`;
+  const inputProps = {
+    id,
+    name,
+    disabled,
+    type: 'checkbox',
+    checked: values.includes(option.value),
+    onChange: handleChange({ value: option.value, values, onChange }),
+    'aria-invalid': ariaInvalid,
+    'aria-describedby': ariaDescribedby,
+  };
 
   if (variant) {
     return (
-      <label data-test={'field-checkbox'} className={'check'} htmlFor={id}>
-        <input id={id} name={input.name} className={'hidden-up pos-abs top-0 left-0'} {...inputProps} />
-        <span className={`btn btn-md btn-block color-white--checked bg-${variant}--checked rounded-0`}>{label}</span>
+      <label data-test={'field--checkbox'} className={'check'} htmlFor={id}>
+        <input className={'hidden-up pos-abs top-0 left-0'} {...inputProps} />
+        <span className={`btn btn-md btn-block color-white--checked bg-${variant}--checked rounded-0`}>
+          {option.label}
+        </span>
       </label>
     );
   }
 
   return (
-    <label data-test={'field-checkbox'} className={'form-check-inline p-l-md m-l-0 m-r-md'} htmlFor={id}>
-      <input id={id} name={input.name} className={'hidden-up pos-abs'} {...inputProps} />
-      <div className={'check-icon-checkbox'}>
+    <label data-test={'field--checkbox'} className={'form-check-inline p-l-md m-l-0 m-r-md'} htmlFor={id}>
+      <input className={'hidden-up pos-abs'} {...inputProps} />
+      <div className={'check-icon-checkbox'} aria-hidden>
         <i className={'ca-gov-icon-check-mark'} />
       </div>
-      {label}
+      {option.label}
     </label>
   );
 };
 
 Checkbox.propTypes = {
+  /** Name of field */
+  name: T.string.isRequired,
+
+  /** Value of checkbox */
+  value: T.oneOfType([
+    T.string,
+    T.array,
+  ]).isRequired,
+
+  /** Option to select */
+  option: propUtils.option.isRequired,
+
   /** Use style variant */
   variant: T.oneOf([
     '',
     'primary',
   ]),
 
-  /** Input from redux-form's Field, attaches name, value, etc  */
-  input: T.object.isRequired,
-
-  /** Value of checkbox */
-  value: T.string.isRequired,
-
-  /** Label to render for this checkbox */
-  label: T.string.isRequired,
-
   /** Disable the input */
   disabled: T.bool,
+
+  /** Label to render for this checkbox */
+  label: T.string,
+
+  /** Called when checkbox changes */
+  onChange: T.func.isRequired,
+
+  /** Accessible indicator for errors existing */
+  'aria-invalid': T.string.isRequired,
+
+  /** Accessible indicator of related information */
+  'aria-describedby': T.string,
 };
 
 Checkbox.defaultProps = {
   variant: '',
   disabled: false,
+  label: null,
+  'aria-describedby': null,
 };
 
 export default Checkbox;