+++ src/lib/components/FieldDate/FieldDate.d.ts
@@ -3,15 +3,18 @@ import * as ReduxForm from 'redux-form';
 
 import * as types from '../../utils/types';
 
-export interface FieldDateProps {
-  /** Input from redux-form's Field, attaches name, value, etc */
-  input: ReduxForm.WrappedFieldInputProps;
+export interface FieldDateProps extends HTMLInputElement {
+  /** Value of date selected */
+  value?: String;
 
   /** Minimum date able to select */
   minDate?: string | Date;
 
   /** Disable the input */
   disabled?: boolean;
+
+  /** Called when date changes */
+  onChange?: Function;
 }
 
 declare class FieldDate extends React.Component<FieldDateProps, {}> {}