+++ src/lib/components/FieldDate/FieldDate.test.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import DatePicker from './DatePicker';
 import { FieldDate } from './FieldDate';
 
 const { testUtils } = global;
@@ -15,16 +14,30 @@ const props = {
 };
 
 let wrapper;
+let instance;
 describe('FieldDate', () => {
   beforeEach(() => {
     wrapper = shallow(<FieldDate {...props} />);
+    instance = wrapper.instance();
   });
 
   it('should render correctly', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('should render a `DatePicker`', () => {
-    expect(wrapper.find(DatePicker).length).toBe(1);
+  it('should format and update the displayed date when a new value is recieved', () => {
+    const input = { ...props.input, value: new Date('2011-09-01T08:00:00.000Z') };
+    instance.componentWillReceiveProps({ ...props, input });
+    expect(wrapper.state('displayText')).toBe('2011-09-01');
+  });
+
+  it('should hide the calender icon when disabled', () => {
+    wrapper.setProps({ disabled: true });
+    expect(wrapper.find(Icon).length).toBe(0);
+  });
+
+  it('should handle invalid dates', () => {
+    const date = instance.getDate('invalid date');
+    expect(date).toBe('');
   });
 });