+++ src/lib/components/FieldDate/index.js
@@ -1,2 +1,1 @@
-export { default as DatePicker } from './DatePicker';
 export { default } from './FieldDate';