+++ src/lib/components/FieldRadioButtons/FieldRadioButtons.d.ts
@@ -3,19 +3,16 @@ import * as ReduxForm from 'redux-form';
 
 import * as types from '../../utils/types';
 
-export interface FieldRadioButtonsProps {
+export interface FieldRadioButtonsProps extends HTMLInputElement {
   /** Use style variant */
   variant?: ''
     | 'primary';
 
-  /** Input from redux-form's Field, attaches name, value, etc */
-  input: ReduxForm.WrappedFieldInputProps;
-
   /** Options to select from */
   options: Array<types.Option>;
 
-  /** Disable the input */
-  disabled?: boolean;
+  /** Label to display to radio button group */
+  label: String;
 }
 
 declare class FieldRadioButtons extends React.Component<FieldRadioButtonsProps, {}> {}