+++ src/lib/components/FieldRadioButtons/FieldRadioButtons.js
@@ -2,38 +2,38 @@ import React from 'react';
 import T from 'prop-types';
 
 import { isEmptyRadio } from '../../utils/validate';
-import withField from '../../utils/withField';
+import * as propUtils from '../../utils/propTypes';
+import { withField } from '../Field';
 
-import FieldWrapper from '../FieldWrapper';
 import RadioButton from './RadioButton';
 
-// group of radio buttons with optional help text and label
-// displays error if validation fails
+// group of radio buttons
 export const FieldRadioButtons = (props) => {
   const {
-    input, options, variant, disabled, ...rest
+    options,
+    label,
+    ...rest
   } = props;
-  const { name } = input;
 
   return (
-    <FieldWrapper data-test={'field-radio-buttons'} name={name} disabled={disabled} {...rest}>
-      <div className={'field-radio'}>
-        {
-          options.filter(x => !x.hidden).map(option => (
-            <RadioButton
-              input={input}
-              key={option.value}
-              variant={variant}
-              disabled={disabled}
-              {...option}
-            />
-          ))
-        }
-      </div>
-    </FieldWrapper>
+    <div
+      data-test={'field--radio-buttons'}
+      className={'field-radio'}
+      role={'group'}
+      aria-label={label}
+    >
+      {
+        options.filter(x => !x.hidden).map(option => (
+          <RadioButton
+            key={option.value}
+            option={option}
+            {...rest}
+          />
+        ))
+      }
+    </div>
   );
 };
-
 
 FieldRadioButtons.propTypes = {
   /** Use style variant */
@@ -42,25 +42,15 @@ FieldRadioButtons.propTypes = {
     'highlight',
   ]),
 
-  /** Input from redux-form's Field, attaches name, value, etc */
-  input: T.object.isRequired,
-
   /** Options to select from */
-  options: T.arrayOf(
-    T.shape({
-      label: T.string.isRequired,
-      value: T.string.isRequired,
-    }),
-  ).isRequired,
+  options: T.arrayOf(propUtils.option).isRequired,
 
-  /** Disable the input */
-  disabled: T.bool,
+  /** Label to display to radio button group */
+  label: T.string.isRequired,
 };
 
 FieldRadioButtons.defaultProps = {
   variant: '',
-  disabled: false,
 };
 
-const withReduxField = withField(isEmptyRadio);
-export default withReduxField(FieldRadioButtons);
+export default withField(isEmptyRadio)(FieldRadioButtons);