+++ src/lib/components/FieldSelect/FieldSelect.js
@@ -3,22 +3,21 @@ import T from 'prop-types';
 import classNames from 'classnames';
 
 import { isEmptyRadio } from '../../utils/validate';
-import withField from '../../utils/withField';
-
-import FieldWrapper from '../FieldWrapper';
+import * as propUtils from '../../utils/propTypes';
+import { withField } from '../Field';
 
 // select with optional help text and label
 // displays error if validation fails
 export const FieldSelect = (props) => {
   const {
-    input, options, disabled, multiple, ...rest
+    value,
+    options,
+    multiple,
+    ...rest
   } = props;
 
-  if (multiple) {
-    // expects array as default if multiple is used
-    input.value = input.value || [];
-  }
-
+  // expects array as default if multiple is used
+  const mappedValue = multiple && !value ? [] : value;
   const cn = classNames([
     'select',
     'field-select',
@@ -26,44 +25,33 @@ export const FieldSelect = (props) => {
   ]);
 
   return (
-    <FieldWrapper data-test={'field-select'} name={input.name} disabled={disabled} {...rest}>
-      <div className={cn}>
-        <select id={input.name} multiple={multiple} disabled={disabled} {...input}>
-          <option disabled hidden style={{ display: 'none' }} value={''} />
-          {
-            options.filter(x => !x.hidden).map(option => (
-              <option key={option.value} value={option.value}>{option.label}</option>
-            ))
-          }
-        </select>
-      </div>
-    </FieldWrapper>
+    <div data-test={'field--select'} className={cn}>
+      <select multiple={multiple} value={mappedValue} {...rest}>
+        <option disabled hidden style={{ display: 'none' }} value={''} />
+        {
+          options.filter(x => !x.hidden).map(option => (
+            <option key={option.value} value={option.value}>{option.label}</option>
+          ))
+        }
+      </select>
+    </div>
   );
 };
 
 FieldSelect.propTypes = {
-  /** Input from redux-form's Field, attaches name, value, etc */
-  input: T.object.isRequired,
+  /** Value of option */
+  value: T.any,
 
   /** Options to select from */
-  options: T.arrayOf(
-    T.shape({
-      label: T.string.isRequired,
-      value: T.string.isRequired,
-    }),
-  ).isRequired,
-
-  /** Disable the input */
-  disabled: T.bool,
+  options: T.arrayOf(propUtils.option).isRequired,
 
   /** Allow multiple selecting */
   multiple: T.bool,
 };
 
 FieldSelect.defaultProps = {
-  disabled: false,
+  value: null,
   multiple: false,
 };
 
-const withReduxField = withField(isEmptyRadio);
-export default withReduxField(FieldSelect);
+export default withField(isEmptyRadio)(FieldSelect);