+++ src/lib/components/FieldTime/index.js
@@ -1,2 +1,0 @@
-export { default } from './FieldTime';
-export { default as TimePicker } from './TimePicker';